force to write 32 bit integer and 64bit float

+++ lib/gsiCovarianceFile.py
@@ -39,8 +39,8 @@ class gsiCovarianceFile:
         f = FortranFile( self.fname, 'w' )
         nchan = len(self.channelIdx)
         f.write_record( np.array([nchan, 8], dtype=np.int32) )
-        f.write_record(self.channelIdx)
-        f.write_record(self.R.T)
+        f.write_record(self.channelIdx.astype(np.int32))
+        f.write_record(self.R.T.astype(np.float64))
         f.close()
 """
 iasi = gsiCovarianceFile('iasi_metop_sea_rcov.bin')
